+++ config/environments/test.rb
@@ -245,7 +245,7 @@ Markus::Application.configure do
   config.action_mailer.delivery_method = :test
   config.action_mailer.asset_host = 'http://localhost:3000'
   config.action_mailer.default_url_options = {host: 'localhost:3000'}
-
+  config.action_mailer.perform_deliveries = true
 
   ###################################################################
   # Resque queues