+++ db/data/autotest_files/custom/script_files/autotest_01.sh
@@ -1,3 +1,3 @@
 #!/bin/bash
 
-python submission.py $0
+python3 submission.py $0